+++ src/application/dto.ts
@@ -1,0 +1,21 @@
+import { User } from "../domen/models";
+
+export class UserDto {
+  id: string;
+  name: string;
+  email: string;
+  role: string;
+  isVerified: boolean;
+
+  constructor(id: string, name: string, email: string, role: string, isVerified: boolean) {
+    this.id = id;
+    this.name = name;
+    this.email = email;
+    this.role = role;
+    this.isVerified = isVerified;
+  }
+
+  static fromDomain(user: User): UserDto {
+    return new UserDto(user.id, user.name, user.email, user.role, user.isVerified);
+  }
+}
