+++ src/controllers/schemas/user.ts
@@ -1,0 +1,23 @@
+import { IsEmail, IsString, MinLength } from 'class-validator';
+
+export class CreateUserSchema {
+  @IsString()
+  name!: string;
+
+  @IsEmail()
+  email!: string;
+
+  @IsString()
+  @MinLength(6)
+  password!: string;
+}
+
+export class LoginUserSchema {
+  @IsEmail()
+  email!: string;
+
+  @IsString()
+  @MinLength(6)
+  password!: string;
+}
+
