--- src/email/application/mail.adapter.interface.ts
@@ -1,3 +1,0 @@
-export interface IMailService {
-  sendMail(to: string, subject: string, html: string): Promise<void>;
-}
