--- src/security/password.adapter.interface.ts
@@ -1,4 +1,0 @@
-export interface IPasswordAdapter {
-  hash(plain: string): Promise<string>;
-  verify(hash: string, plain: string): Promise<boolean>;
-}
