--- src/users/application/users.repository.interface.ts
@@ -1,7 +1,0 @@
-import { UserEntity } from '../infrastructure/user.entity';
-
-export interface IUsersRepository {
-  findAll(): Promise<UserEntity[]>;
-  insert(id: string, name: string, email: string, hashedPassword: string): Promise<UserEntity>;
-  findPasswordByEmail(email: string): Promise<string | null>;
-}
