--- src/users/controllers/schemas/create-user.schema.ts
@@ -1,14 +1,0 @@
-import { IsEmail, IsString, MinLength } from 'class-validator';
-
-export class CreateUserSchema {
-  @IsString()
-  name!: string;
-
-  @IsEmail()
-  email!: string;
-
-  @IsString()
-  @MinLength(6)
-  password!: string;
-}
-
