--- src/users/users.module.ts
@@ -1,21 +1,0 @@
-// src/users/users.module.ts
-import { Module } from '@nestjs/common';
-import { UsersController, AuthController } from './controllers/users.controller';
-import { UsersService } from './application/users.service';
-import { UsersRepository } from './infrastructure/users.repository';
-import { SecurityModule } from '../security/security.module';
-import { EmailModule } from '../email/email.module'; 
-
-@Module({
-  imports: [SecurityModule, EmailModule],
-  controllers: [UsersController, AuthController],
-  providers: [
-    UsersService,
-    {
-      provide: 'IUsersRepository',
-      useClass: UsersRepository,
-    },
-  ],
-  exports: [UsersService],
-})
-export class UsersModule {}
